Add a pseudo-module, one which maintains the namespace of it's parent module

The usage of a pseudo-module is nearly identical to the module, but inputs & outputs in the pseudo-module are shared between the parent and the pseudo-module. It's usefulness shines when developing numerous interacting inputs, say for a complex plot, and those inputs need be modular but also must interact with one another in a confluence fashion to determine how the plot renders. The pseudo-module allows developers to functionalize input components whose values must necessarily must be considered in concert with one another. Using modules, the developer must keep track of shared inputs explicitly between modules using reactiveVal and/or reactiveValues and pass them as arguments to any nested modules. As a codebase changes, and inputs are added and removed, keeping track of where confluences of interactions are stored and updating arguments accordingly is cumulative technical debt. With pseudo-modules shared reactive inputs allow the developer to logically segment interactivity involving confluences of inputs as they see fit without incurring the technical debt of passing reactiveVal and/or reactiveValues around when interactions are removed or added.

Usage

add_pseudo_module(
  name,
  path = "R",
  export = FALSE,
  pm_name = sprintf("pmod_%s", name),
  pm_ui = sprintf("pmod_%s_ui", name),
  pm_server = sprintf("pmod_%s_server", name),
  use_ud = TRUE,
  open = TRUE
)

Arguments

name

chr name of the module

path

chr directory in which to create it

export

lgl whether it should be exported in the package

pm_name

chr The name of file

pm_ui

chr The name of the module ui function

pm_server

chr The name of the module server function

use_ud

lgl Whether to include the session$userData environment as a shorthand variable ud in the module for ease of access to a user's session state variables.

Value

file The new file opens