Create a row with columns that dynamically resize Must use div(class = 'column', ...) or col_auto() for internal columns to function properly.

Create a row with columns that dynamically resize Must use div(class = 'column', ...) or col_auto() for internal columns to function properly.

Usage

dynamic_row(...)

Arguments

...

Elements to include within the page

Value

A UI definition that can be passed to the shinyUI function.

See also

Examples

dynamic_row(col_auto(bs4Dash::box(width = 12)), col_auto(bs4Dash::box(width = 12)))
#> <div class="row" style="display: grid; grid-template-columns: 1 fr 1 fr; grid-template-areas: &quot;a a&quot;;">
#>   <div class="column">
#>     <div class="col-sm-12">
#>       <div class="card bs4Dash ">
#>         <div class="card-header">
#>           <h3 class="card-title">‌</h3>
#>           <div class="card-tools float-right">
#>             <button class="btn btn-tool btn-sm" type="button" data-card-widget="collapse">
#>               <i class="fas fa-minus" role="presentation" aria-label="minus icon"></i>
#>             </button>
#>           </div>
#>         </div>
#>         <div class="card-body"></div>
#>       </div>
#>       <script type="application/json">{"solidHeader":true,"width":12,"collapsible":true,"closable":false,"maximizable":false,"gradient":false}</script>
#>     </div>
#>   </div>
#>   <div class="column">
#>     <div class="col-sm-12">
#>       <div class="card bs4Dash ">
#>         <div class="card-header">
#>           <h3 class="card-title">‌</h3>
#>           <div class="card-tools float-right">
#>             <button class="btn btn-tool btn-sm" type="button" data-card-widget="collapse">
#>               <i class="fas fa-minus" role="presentation" aria-label="minus icon"></i>
#>             </button>
#>           </div>
#>         </div>
#>         <div class="card-body"></div>
#>       </div>
#>       <script type="application/json">{"solidHeader":true,"width":12,"collapsible":true,"closable":false,"maximizable":false,"gradient":false}</script>
#>     </div>
#>   </div>
#> </div>