Create a driver.js callout Must include shinyVirga::use_driver.js() & useShinyJS in the head of the page.

Create a driver.js callout Must include shinyVirga::use_driver.js() & useShinyJS in the head of the page.

Usage

js_callout(
  el,
  title,
  description,
  side = NULL,
  align = NULL,
  animate = TRUE,
  overlayColor = "black",
  smoothScroll = FALSE,
  allowClose = TRUE,
  overlayOpacity = 0.5,
  stagePadding = 10,
  stageRadius = 5,
  allowKeyboardControl = TRUE,
  disableActiveInteraction = FALSE,
  popoverClass = "scoped-class",
  popoverOffset = 10,
  showButtons = NULL,
  disableButtons = NULL,
  showProgress = FALSE,
  progressText = FALSE,
  nextBtnText = "Next",
  prevBtnText = "Previous",
  doneBtnText = "Done",
  onPopoverRender = NULL,
  onHighlightStarted = NULL,
  onHighlighted = NULL,
  onDeselected = NULL,
  onDestroyStarted = NULL,
  onDestroyed = NULL,
  onNextClick = NULL,
  onPrevClick = NULL,
  onCloseClick = NULL,
  asis = FALSE,
  .ns = ns_find()
)

Arguments

el

chr ID or class of element. If a character not prefixed by . for a class or # for an ID, it will be assumed and ID and # will be added.

title

chr Title of popover

description

chr Description for it

side

chr The position of the popover relative to the target element. One of "top", "right", "bottom", "left".

align

chr The alignment of the popover relative to the target element. One of "start", 'center', "end".

animate

lgl Whether to animate the product tour. (default: true)

overlayColor

chr Overlay color. (default: black) This is useful when you have a dark background and want to highlight elements with a light background color.

smoothScroll

lgl Whether to smooth scroll to the highlighted element. (default: false)

allowClose

lgl Whether to allow closing the popover by clicking on the backdrop. (default: true)

overlayOpacity

num Opacity of the backdrop. (default: 0.5)

stagePadding

num Distance between the highlighted element and the cutout. (default: 10)

stageRadius

num Radius of the cutout around the highlighted element. (default: 5)

allowKeyboardControl

lgl Whether to allow keyboard navigation. (default: true)

disableActiveInteraction

lgl Whether to disable interaction with the highlighted element. (default: false)

popoverClass

chr If you want to add custom class to the popover

popoverOffset

num Distance between the popover and the highlighted element. (default: 10)

showButtons

chr Array of buttons to show in the popover. Defaults to "next", "previous", "close" for product tours and [] for single element highlighting.

disableButtons

chr Array of buttons to disable. This is useful when you want to show some of the buttons, but disable some of them.

showProgress

lgl Whether to show the progress text in popover. (default: false)

progressText

chr See docs for details

nextBtnText

chr See docs for details

prevBtnText

chr See docs for details

doneBtnText

chr See docs for details

onPopoverRender

chr See docs for details

onHighlightStarted

chr See docs for details

onHighlighted

chr See docs for details

onDeselected

chr See docs for details

onDestroyStarted

chr See docs for details

onDestroyed

chr See docs for details

onNextClick

chr See docs for details

onPrevClick

chr See docs for details

onCloseClick

chr See docs for details

asis

If TRUE, use the ID as-is even when inside a module (instead of adding the namespace prefix to the ID).

.ns

fun ns function. Typically found automatically.

closeBtnText

chr See docs for details

See also

Examples

js_callout('my_id', "Tooltip title", "Tooltip description")
#> Warning: Could not find ns function. Using `function(x) x`
#> Error: shinyjs: could not find the Shiny session object. This usually happens when a shinyjs function is called from a context that wasn't set up by a Shiny session.