Add Transifex Language selector

Must set up an environment variable entitled TRANSIFEX_API_KEY with your API Key created here and run transifex_select_server at least once in your app.

Usage

mod_transifex_select(
  inputId,
  label,
  choices = c(English = "en", Spanish = "es"),
  selected = c(English = "en"),
  multiple = FALSE,
  options = list(),
  choicesOpt = NULL,
  width = NULL,
  class = NULL,
  inline = FALSE,
  include_deps = FALSE,
  key = Sys.getenv("TRANSIFEX_API_KEY", ""),
  ...
)

Arguments

inputId

The input slot that will be used to access the value.

label

Display label for the control, or NULL for no label.

choices

List of values to select from. If elements of the list are named then that name rather than the value is displayed to the user.

selected

The initially selected value (or multiple values if multiple = TRUE). If not specified then defaults to the first value for single-select lists and no values for multiple select lists.

multiple

Is selection of multiple items allowed?

options

List of options, see pickerOptions for all available options. To limit the number of selection possible, see example below.

choicesOpt

Options for choices in the dropdown menu.

width

The width of the input : 'auto', 'fit', '100px', '75%'.

inline

Display picker inline, to have label and menu on same line use width = "fit".

...

Arguments passed on to mod_transifex_deps

transifex_options

list List of options to be passed to the liveSettings object. The option manual_init must be TRUE when including the dependencies anywhere other than in the head of the application.

Value

shiny.tag

Details

Implementation adapted from the Transifex docs

See also