A default full width row box.

A default full width row box.

Usage

ui_row(
  ...,
  title = NULL,
  footer = NULL,
  status = NULL,
  solidHeader = FALSE,
  background = NULL,
  width = 12,
  height = NULL,
  collapsible = TRUE,
  collapsed = FALSE,
  closable = FALSE,
  maximizable = FALSE,
  icon = NULL,
  tip_icon = NULL,
  gradient = FALSE,
  boxToolSize = "sm",
  elevation = NULL,
  headerBorder = TRUE,
  label = NULL,
  dropdownMenu = NULL,
  sidebar = NULL,
  id = NULL,
  box = TRUE,
  row = TRUE,
  class = NULL,
  add_attribs = NULL,
  state_input = !is.null(id)
)

Arguments

...

Contents of the box.

title

Optional title.

footer

Optional footer text.

status

The status of the item. This determines the item's background color. Valid statuses are defined as follows:

primary: #007bff .

secondary: #6c757d .

info: #17a2b8 .

success: #28a745 .

warning: #ffc107 .

danger: #dc3545 .

gray-dark: #343a40 .

gray: #adb5bd .

white: #fff .

indigo: #6610f2 .

lightblue: #3c8dbc .

navy: #001f3f .

purple: #605ca8 .

fuchsia: #f012be .

pink: #e83e8c .

maroon: #d81b60 .

orange: #ff851b .

lime: #01ff70 .

teal: #39cccc .

olive: #3d9970 .

solidHeader

Should the header be shown with a solid color background?

background

If NULL (the default), the background of the box will be white. Otherwise, a color string. Valid colors are listed in validColors. See below:

primary: #007bff .

secondary: #6c757d .

info: #17a2b8 .

success: #28a745 .

warning: #ffc107 .

danger: #dc3545 .

gray-dark: #343a40 .

gray: #adb5bd .

white: #fff .

indigo: #6610f2 .

lightblue: #3c8dbc .

navy: #001f3f .

purple: #605ca8 .

fuchsia: #f012be .

pink: #e83e8c .

maroon: #d81b60 .

orange: #ff851b .

lime: #01ff70 .

teal: #39cccc .

olive: #3d9970 .

width

The width of the box, using the Bootstrap grid system. This is used for row-based layouts. The overall width of a region is 12, so the default card width of 6 occupies 1/2 of that width. For column-based layouts, use NULL for the width; the width is set by the column that contains the box.

height

The height of a box, in pixels or other CSS unit. By default the height scales automatically with the content.

collapsible

If TRUE, display a button in the upper right that allows the user to collapse the box.

collapsed

If TRUE, start collapsed. This must be used with collapsible=TRUE.

closable

If TRUE, display a button in the upper right that allows the user to close the box.

maximizable

If TRUE, the card can be displayed in full screen mode.

icon

Header icon. Displayed before title. Expect icon.

tip_icon

Tip icon. Tooltip Icon displayed after title. Expect tippy.

gradient

Whether to allow gradient effect for the background color. Default to FALSE.

boxToolSize

Size of the toolbox: choose among "xs", "sm", "md", "lg".

elevation

Card elevation.

headerBorder

Whether to display a border between the header and body. TRUE by default

label

Slot for boxLabel.

dropdownMenu

List of items in the boxtool dropdown menu. Use boxDropdown.

sidebar

Slot for boxSidebar.

id

Card id.

box

lgl Whether to box the contents or just put them in row. Default TRUE

row

lgl Whether to wrap the output in a row. Default TRUE

class

chr A class to add to the top level div.row

add_attribs

list of attributes to append to arbitrary tags in the format list([jQuery selector] = list([attribute] = [value]))

state_input

lgl Whether to make a logical input value with the ID provided. Must ensure id is namespaced with the ns function if calling inside a module. The resulting input will be input[['{id}_state']] where id is the value passed to the id argument.

Value

A full-width box nested in a row

See also

Examples

ui_row(tags$p("Hi"))
#> <div class="row ui_row">
#>   <div class="col-sm-12">
#>     <div class="card bs4Dash ">
#>       <div class="card-header">
#>         <h3 class="card-title">‌</h3>
#>         <div class="card-tools float-right">
#>           <button class="btn btn-tool btn-sm" type="button" data-card-widget="collapse">
#>             <i class="fas fa-minus" role="presentation" aria-label="minus icon"></i>
#>           </button>
#>         </div>
#>       </div>
#>       <div class="card-body">
#>         <p>Hi</p>
#>       </div>
#>     </div>
#>     <script type="application/json">{"solidHeader":true,"width":12,"collapsible":true,"closable":false,"maximizable":false,"gradient":false}</script>
#>   </div>
#> </div>